move appmgmt to top level pkg
