+++ pkg/appmgmt/sparkoperator/spark.go
